% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_binned_checkerplot.R
\name{binnedcp.plot}
\alias{binnedcp.plot}
\title{Binned checker plot to display results from all-by-all screens}
\usage{
binnedcp.plot(ppi_prediction_result, fraction_na = 0.05)
}
\arguments{
\item{ppi_prediction_result:}{result object from the ppi.prediction() function.}

\item{fraction_na:}{fraction of interactions allowed to be \emph{NA} to still draw the checker plot}
}
\value{
ggplot
}
\description{
Binned checker plot to display results from all-by-all screens
}
\examples{
data("example_ppi_prediction")
binnedhm.plot(example_ppi_prediction)
}
