% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion.matrix}
\alias{confusion.matrix}
\title{Function to calculate a confusion matrix for}
\usage{
confusion.matrix(ppi_prediction_result)
}
\arguments{
\item{ppi_prediction_result}{}
}
\description{
Function to calculate a confusion matrix for
}
\examples{
data("example_ppi_prediction")
confusion_matrix <- confusion.matrix(example_ppi_prediction)

}
