% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_ppi_prediction.R
\docType{data}
\name{example_ppi_prediction}
\alias{example_ppi_prediction}
\title{Example prediction}
\format{
Contains two object: \code{example_ppi_prediction} from \code{ppi.prediction(PPIdf = luthy_reference_sets, referenceSet = luthy_reference_sets)} and \code{example_learning_curve} from \code{learning.curve(example_ppi_prediction)}
}
\source{
https://doi.org/10.1101/2023.06.14.544560
}
\usage{
data("example_ppi_prediction")
}
\description{
Example prediction
}
\keyword{datasets}
