% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luthy_reference_sets.R
\docType{data}
\name{luthy_reference_sets}
\alias{luthy_reference_sets}
\title{LuTHy reference sets}
\format{
A data frame containing the LuTHy-BRET and LuTHy-LuC results from the PRS-v2 and hsRRS-v2
}
\source{
https://doi.org/10.1101/2023.06.14.544560
}
\usage{
data("luthy_reference_sets")
}
\description{
LuTHy reference sets
}
\keyword{datasets}
