% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{make.grid.BRET}
\alias{make.grid.BRET}
\title{Generate a grid to plot for BRET predictions}
\usage{
make.grid.BRET(x, n = 100)
}
\arguments{
\item{x:}{takes as input the from result of ppi.prediction()$testMat}

\item{n:}{size of the grid}
}
\description{
Generate a grid to plot for BRET predictions
}
