% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot.probDis}
\alias{plot.probDis}
\title{Plot a scatter plot of the probability distribution against the primary assay feature}
\usage{
\method{plot}{probDis}(ppi_prediction_result)
}
\arguments{
\item{ppi_prediction_result}{}
}
\description{
Plot a scatter plot of the probability distribution against the primary assay feature
}
