% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot.probGrid.BRET}
\alias{plot.probGrid.BRET}
\title{Plot a probability grid from the mean probabilities from the 'ensembleSize' number of models for the LuTHy-BRET assay}
\usage{
\method{plot}{probGrid.BRET}(ppi_prediction_result)
}
\arguments{
\item{ppi_prediction_result}{}
}
\description{
Plot a probability grid from the mean probabilities from the 'ensembleSize' number of models for the LuTHy-BRET assay
}
