% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recovery.R
\name{plot.recovery}
\alias{plot.recovery}
\title{Plot a bar diagram of the recovery rates at 50\%, 75\% and 95\% probability}
\usage{
\method{plot}{recovery}(ppi_prediction_result)
}
\arguments{
\item{ppi_prediction_result}{}
}
\description{
Plot a bar diagram of the recovery rates at 50\%, 75\% and 95\% probability
}
