% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot.roc}
\alias{plot.roc}
\title{Plot ROC curve of the predicted probabilities against the assay features the ML was trained on}
\usage{
\method{plot}{roc}(ppi_prediction_result)
}
\arguments{
\item{ppi_prediction_result}{}
}
\description{
Plot ROC curve of the predicted probabilities against the assay features the ML was trained on
}
