% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_probGrid.R
\name{probGrid.plot}
\alias{probGrid.plot}
\title{Plot a probability grid from the mean probabilities from the 'ensembleSize' number of models}
\usage{
probGrid.plot(
  ppi_prediction_result,
  n = 100,
  x.log.scale = TRUE,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{ylim}{}
}
\description{
Plot a probability grid from the mean probabilities from the 'ensembleSize' number of models
}
