% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recovery_rate.R
\name{recovery.rate}
\alias{recovery.rate}
\title{Function to calculate the recovery rate from binary PPI assays.}
\usage{
recovery.rate(
  referenceSet = NULL,
  PPIdf = NULL,
  assay = "mean_cBRET",
  construct.scaling = NULL,
  range = c(0.25, 0.75),
  iter.scaler = TRUE,
  cutoff = "all",
  fns = "median",
  n.sd = 1,
  n.CI = 0.99,
  negative.reference = "RRS",
  complex.parameters = NULL,
  verbose = TRUE
)
}
\arguments{
\item{referenceSet:}{reference PPI data set used to determine cutoffs to calculate recovery rates}

\item{PPIdf:}{optional binary PPI data set containing interactions classified using the cutoffs from the reference sets}

\item{assay:}{assay in the referenceSet and PPIdf to calculate recovery rates}

\item{construct.scaling:}{scaling of constructs by "median.normalization" or "robust.scaler"}

\item{range:}{probs in stats::quantile(x, ...): numeric vector of probabilities}

\item{iter.scaler:}{if TRUE and when using "robust.scaler" it iteratively performs robust scaler normalization until the IQR of each construct is within the IQR of all loaded data sets}

\item{cutoff:}{cutoffs are calculated based on "all" interactions, or for each "construct" or for each "orientation"}

\item{fns:}{function used to calculate the cutoffs: "median", "mean", "CI" or "max"}

\item{n.sd:}{number of sd used to calculate the cutoffs for functions "median" and "mean"}

\item{n.CI:}{confidence intervall used to calculate the cutoff for function "CI"}

\item{negative.reference:}{string in the column "complex" to specify the negative/random interactions}

\item{complex.parameters:}{additional grouping parameters from which the recovery rates are calculated}

\item{verbose:}{give detailed information}
}
\description{
Function to calculate the recovery rate from binary PPI assays.
}
