% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{roc.plot}
\alias{roc.plot}
\title{Plot ROC curve of the predicted probabilities against the assay features the ML was trained on}
\usage{
roc.plot(ppi_prediction_result)
}
\arguments{
\item{ppi_prediction_result:}{result object from the ppi.prediction() function.}
}
\value{
a ggplot2 object
}
\description{
Plot ROC curve of the predicted probabilities against the assay features the ML was trained on
}
