% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_curve.R
\name{learning.curve}
\alias{learning.curve}
\title{Function to calculate and plot the accuracy, hinge loss and binary cross-entropy loss of a machine learning prediction using the ppi.prediction function.}
\usage{
learning.curve(
  ppi_prediction_result,
  train_sizes = base::seq(0.1, 1, by = 0.1),
  models = "all",
  verbose = TRUE
)
}
\arguments{
\item{ppi_prediction_result:}{result from the function ppi.prediction}

\item{train_sizes:}{base::sequence of fraction of training sizes to be used for calculation from >0 to 1}

\item{models:}{Integer of models used to calculate the loss functions. If "all" then all models as specified by the ensembleSize in ppi.prediction will be used.}

\item{verbose:}{boolean, prints detailed informations}
}
\description{
Function to calculate and plot the accuracy, hinge loss and binary cross-entropy loss of a machine learning prediction using the ppi.prediction function.
}
