% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_probDis.R
\name{probDis.plot}
\alias{probDis.plot}
\title{Plot a scatter plot of the probability distribution against the primary assay feature}
\usage{
probDis.plot(ppi_prediction_result)
}
\arguments{
\item{ppi_prediction_result:}{result object from the ppi.prediction() function.}
}
\value{
a ggplot2 object
}
\description{
Plot a scatter plot of the probability distribution against the primary assay feature
}
