% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_porbDis.R
\name{probDis.plot}
\alias{probDis.plot}
\title{Plot a scatter plot of the probability distribution against the primary assay feature}
\usage{
probDis.plot(ppi_prediction_result)
}
\arguments{
\item{ppi_prediction_result}{}
}
\description{
Plot a scatter plot of the probability distribution against the primary assay feature
}
