% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recovery.R
\name{recovery.plot}
\alias{recovery.plot}
\title{Plot a bar diagram of the recovery rates at 50\%, 75\% and 95\% probability}
\usage{
recovery.plot(ppi_prediction_result)
}
\arguments{
\item{ppi_prediction_result}{}
}
\description{
Plot a bar diagram of the recovery rates at 50\%, 75\% and 95\% probability
}
